third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/script-loads-with-img-src-expected.txt

CONSOLE ERROR: Unrecognized Content-Security-Policy directive 'script-img'.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
